djangoapp/requirements.txt
Django>=4.2.1,<4.3
psycopg2-binary>=2.9.6,<2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==24.3
dj-rest-auth==7.0.1
django-allauth==65.3.1